requirements.txt
beautifulsoup4==4.12.3
google-generativeai==0.8.2
gtts==2.5.3
pymupdf==1.24.11
python-dotenv==1.0.1